requirements.txt
pandas
requests
beautifulsoup4
lxml
python-dotenv